Extract testing trait
